requirements.txt
requests
PyYAML
click
tabulate
colorama
